+++ node/src/types/kernel/api/notebook/createNotebook/payload.d.ts
@@ -15,14 +15,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-
-/* directories */
-export * as closeNotebook from "./closeNotebook";
-export * as createNotebook from "./createNotebook";
-export * as getNotebookConf from "./getNotebookConf";
-export * as lsNotebooks from "./lsNotebooks";
-export * as openNotebook from "./openNotebook";
-export * as removeNotebook from "./removeNotebook";
-export * as renameNotebook from "./renameNotebook";
-
-/* flles */
+/**
+ * Create a notebook
+ */
+export interface IPayload {
+    /**
+     * notebook name
+     */
+    name: string;
+}